+++ library/src/androidTest/java/com/twilio/video/VideoTopologyParameterizedTest.java
@@ -2,7 +2,6 @@ package com.twilio.video;
 
 import android.support.test.filters.LargeTest;
 import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
 
 import com.twilio.video.base.BaseClientTest;
 import com.twilio.video.helper.CallbackHelper;
@@ -31,7 +30,7 @@ import static org.junit.Assert.assertTrue;
 
 @RunWith(Parameterized.class)
 @LargeTest
-public class VideoClientTopologyParameterizedTest extends BaseClientTest {
+public class VideoTopologyParameterizedTest extends BaseClientTest {
     @Parameterized.Parameters(name = "{0}")
     public static Iterable<Object[]> data() {
         return Arrays.asList(new Object[][]{
@@ -49,7 +48,7 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
     private CallbackHelper.FakeRoomListener roomListener;
     private final Topology topology;
 
-    public VideoClientTopologyParameterizedTest(Topology topology) {
+    public VideoTopologyParameterizedTest(Topology topology) {
         this.topology = topology;
     }
 
@@ -62,7 +61,7 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
         token = CredentialsUtils.getAccessToken(Constants.PARTICIPANT_ALICE, topology);
         roomName = RandUtils.generateRandomString(20);
         localMedia = LocalMedia.create(mediaTestActivity);
-        VideoClient.setLogLevel(LogLevel.ALL);
+        Video.setLogLevel(LogLevel.ALL);
     }
 
     @After
@@ -77,25 +76,25 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
         ConnectOptions connectOptions = new ConnectOptions.Builder(token)
             .roomName(roomName)
             .build();
-        Room room = VideoClient.connect(null, connectOptions, roomListener);
+        Room room = Video.connect(null, connectOptions, roomListener);
     }
 
     @Test(expected = NullPointerException.class)
     public void connect_shouldThrowExceptionWhenConnectOptionsIsNull() {
-        Room room = VideoClient.connect(mediaTestActivity, null, roomListener);
+        Room room = Video.connect(mediaTestActivity, null, roomListener);
     }
 
     @Test
     public void logLevel_shouldBeRetained() {
-        VideoClient.setLogLevel(LogLevel.DEBUG);
-        assertEquals(LogLevel.DEBUG, VideoClient.getLogLevel());
+        Video.setLogLevel(LogLevel.DEBUG);
+        assertEquals(LogLevel.DEBUG, Video.getLogLevel());
     }
 
     @Test
     public void getVersion_shouldReturnValidSemVerFormattedVersion() {
         String semVerRegex = "^([0-9]+)\\.([0-9]+)\\.([0-9]+)(?:-([0-9A-" +
                 "Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+[0-9A-Za-z-]+)?$";
-        String version = VideoClient.getVersion();
+        String version = Video.getVersion();
 
         assertNotNull(version);
         assertTrue(version.matches(semVerRegex));
@@ -106,7 +105,7 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
         roomListener.onConnectedLatch = new CountDownLatch(1);
         ConnectOptions connectOptions = new ConnectOptions.Builder(token)
             .build();
-        Room room = VideoClient.connect(mediaTestActivity, connectOptions, roomListener);
+        Room room = Video.connect(mediaTestActivity, connectOptions, roomListener);
         assertTrue(roomListener.onConnectedLatch.await(20, TimeUnit.SECONDS));
         assertEquals(room.getSid(), room.getName());
         room.disconnect();
@@ -116,7 +115,7 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
     public void connect_shouldThrowExceptionWhenRoomListenerIsNull() {
         ConnectOptions connectOptions = new ConnectOptions.Builder(token)
             .build();
-        VideoClient.connect(mediaTestActivity, connectOptions, null);
+        Video.connect(mediaTestActivity, connectOptions, null);
     }
 
     @Test
@@ -128,7 +127,7 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
                 .roomName(roomName)
                 .localMedia(localMedia)
                 .build();
-        Room room = VideoClient.connect(mediaTestActivity, connectOptions, roomListener);
+        Room room = Video.connect(mediaTestActivity, connectOptions, roomListener);
         assertTrue(roomListener.onConnectedLatch.await(20, TimeUnit.SECONDS));
         assertNotNull(room.getLocalParticipant().getLocalMedia());
         room.disconnect();
@@ -143,7 +142,7 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
                 .roomName(roomName)
                 .localMedia(localMedia)
                 .build();
-        Room room = VideoClient.connect(mediaTestActivity, connectOptions, roomListener);
+        Room room = Video.connect(mediaTestActivity, connectOptions, roomListener);
         assertTrue(roomListener.onConnectedLatch.await(20, TimeUnit.SECONDS));
         assertNotNull(room.getLocalParticipant().getLocalMedia());
         assertTrue(localMedia.removeAudioTrack(localAudioTrack));
@@ -160,7 +159,7 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
                 .roomName(roomName)
                 .localMedia(localMedia)
                 .build();
-        Room room = VideoClient.connect(mediaTestActivity, connectOptions, roomListener);
+        Room room = Video.connect(mediaTestActivity, connectOptions, roomListener);
         assertTrue(roomListener.onConnectedLatch.await(20, TimeUnit.SECONDS));
         assertNotNull(room.getLocalParticipant().getLocalMedia());
         assertTrue(localMedia.removeVideoTrack(localVideoTrack));
@@ -178,7 +177,7 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
                 .roomName(roomName)
                 .localMedia(localMedia)
                 .build();
-        Room room = VideoClient.connect(mediaTestActivity, connectOptions, roomListener);
+        Room room = Video.connect(mediaTestActivity, connectOptions, roomListener);
         assertTrue(roomListener.onConnectedLatch.await(20, TimeUnit.SECONDS));
         assertNotNull(room.getLocalParticipant().getLocalMedia());
         assertTrue(localMedia.removeAudioTrack(localAudioTrack));
@@ -193,7 +192,7 @@ public class VideoClientTopologyParameterizedTest extends BaseClientTest {
             .roomName(roomName)
             .localMedia(localMedia)
             .build();
-        Room room = VideoClient.connect(mediaTestActivity, connectOptions, roomListener);
+        Room room = Video.connect(mediaTestActivity, connectOptions, roomListener);
         assertTrue(roomListener.onConnectFailureLatch.await(20, TimeUnit.SECONDS));
         assertEquals(roomListener.getTwilioException().getCode(),
             TwilioException.ACCESS_TOKEN_INVALID_EXCEPTION);